Implement dbimport -p argument to import only platform-related elements

+++ src/main/java/decodes/dbimport/DbImport.java
@@ -98,15 +98,15 @@ import decodes.sql.DbKey;
 import decodes.sql.PlatformListIO;
 import decodes.sql.SqlDatabaseIO;
 import decodes.tsdb.CompAppInfo;
-import decodes.tsdb.ConstraintException;
 import decodes.tsdb.DbIoException;
 import decodes.util.*;
-import decodes.cwms.CwmsSqlDatabaseIO;
 import decodes.db.*;
 import decodes.xml.DataTypeEquivalenceListParser;
+import decodes.xml.ElementFilter;
 import decodes.xml.EngineeringUnitParser;
 import decodes.xml.UnitConverterParser;
 import decodes.xml.XmlDatabaseIO;
+import decodes.xml.XmlDbTags;
 import decodes.xml.TopLevelParser;
 import decodes.xml.EnumParser;
 
@@ -153,6 +153,9 @@ public class DbImport
 	static BooleanToken allowHistoricalArg = new BooleanToken("H",
 		"Allow import of historical versions. (default=ignore.)", "",
 		TokenOptions.optSwitch, false);
+	static BooleanToken platformRelatedOnlyArg = new BooleanToken("p",
+		"Import ONLY platform-related elements. (default=import all.)", "",
+		TokenOptions.optSwitch, false);
 
 	
 	static
@@ -170,6 +173,7 @@ public class DbImport
 		cmdLineArgs.addToken(overwriteDb);
 		cmdLineArgs.addToken(overwriteDbConfirm);
 		cmdLineArgs.addToken(reflistArg);
+		cmdLineArgs.addToken(platformRelatedOnlyArg);
 	}
 
 	/**
@@ -263,9 +267,7 @@ Logger.instance().debug3("After normalizeTheDb, there are "
 	TopLevelParser topParser; // Top level XML parser.
 	Vector<IdDatabaseObject> newObjects;   // Stores new DatabaseObjects to be added.
 	ArrayList<IdDatabaseObject> toDelete = new ArrayList<IdDatabaseObject>();
-//	boolean writeEnumList;
 	boolean writePlatformList;
-//	boolean writeDataTypes;
 	private Pdt pdt = null;
 	
 	DbImport()
@@ -323,8 +325,10 @@ Logger.instance().debug3("After normalizeTheDb, there are "
 			}
 			if (!overwriteDbConfirm.getValue())
 			{
-				System.out.print("Are you sure you want to replace the "
-					+ "ENTIRE contents of the database at '" + dbloc + "' (y/n)?");
+				System.out.print("Are you sure you want to replace "
+					+ (platformRelatedOnlyArg.getValue() ? "ALL Platform Related Records"
+					   : "ALL DECODES records")
+					+ " from the database at '" + dbloc + "' (y/n)?");
 				String answer = System.console().readLine();
 				if (!answer.toLowerCase().equals("y"))
 					System.exit(1);
@@ -391,68 +395,44 @@ Logger.instance().debug3("After normalizeTheDb, there are "
 		throws DatabaseException
 	{
 		Logger.instance().info("Since the OVERWRITE option was given, deleting current database.");
-		ScheduleEntryDAI scheduleEntryDAO = theDbio.makeScheduleEntryDAO();
-		if (scheduleEntryDAO != null)
+		if (!platformRelatedOnlyArg.getValue())
 		{
-			try
-			{
-				for(ScheduleEntry se : theDb.schedEntryList)
-					scheduleEntryDAO.deleteScheduleEntry(se);
-			}
-			catch (DbIoException ex)
-			{
-				warning("Cannot delete schedule entry: " + ex);
-			}
-			finally
-			{
-				scheduleEntryDAO.close();
-			}
-			theDb.schedEntryList.clear();
-		}
-		for(Iterator<RoutingSpec> rsit = theDb.routingSpecList.iterator(); rsit.hasNext(); )
-		{
-			RoutingSpec rs = rsit.next();
-			theDbio.deleteRoutingSpec(rs);
-		}
-		theDb.routingSpecList.clear();
-		
-		for(Iterator<DataSource> dsit = theDb.dataSourceList.iterator(); dsit.hasNext(); )
-		{
-			DataSource ds = dsit.next();
-			theDbio.deleteDataSource(ds);
-		}
-		theDb.dataSourceList.clear();
-
-		LoadingAppDAI loadingAppDAO = theDbio.makeLoadingAppDAO();
-		try
-		{
-			for(CompAppInfo cai : theDb.loadingAppList)
+			Logger.instance().info("ONLY deleting platform-related entities.");
+			ScheduleEntryDAI scheduleEntryDAO = theDbio.makeScheduleEntryDAO();
+			if (scheduleEntryDAO != null)
 			{
 				try
 				{
-					loadingAppDAO.deleteComputationApp(cai);
+					for(ScheduleEntry se : theDb.schedEntryList)
+						scheduleEntryDAO.deleteScheduleEntry(se);
 				}
-				catch (ConstraintException ex)
+				catch (DbIoException ex)
 				{
-					warning("Cannot delete app '" + cai.getAppName() + "': " + ex);
-				}	
+					warning("Cannot delete schedule entry: " + ex);
+				}
+				finally
+				{
+					scheduleEntryDAO.close();
+				}
+				theDb.schedEntryList.clear();
 			}
-			// Now recreate the 'utility' application which is required by compimport and
-			// others.
-			CompAppInfo utility = new CompAppInfo();
-			utility.setAppName("utility");
-			utility.setProperty("appType", "utility");
-			loadingAppDAO.writeComputationApp(utility);
+			for(Iterator<RoutingSpec> rsit = theDb.routingSpecList.iterator(); rsit.hasNext(); )
+			{
+				RoutingSpec rs = rsit.next();
+				theDbio.deleteRoutingSpec(rs);
+			}
+			theDb.routingSpecList.clear();
+			
+			for(Iterator<DataSource> dsit = theDb.dataSourceList.iterator(); dsit.hasNext(); )
+			{
+				DataSource ds = dsit.next();
+				theDbio.deleteDataSource(ds);
+			}
+			theDb.dataSourceList.clear();
+	
+			// NOTE: Never delete loading apps -- too many dependencies outside of DECODES.
+			// (block of code for removing loading apps removed 10/26/2022)
 		}
-		catch (DbIoException ex)
-		{
-			warning("Cannot delete loading apps: " + ex);
-		}
-		finally
-		{
-			loadingAppDAO.close();
-		}
-		theDb.loadingAppList.clear();
 
 		for(NetworkList nl : theDb.networkListList.getList())
 			theDbio.deleteNetworkList(nl);
@@ -462,16 +442,8 @@ Logger.instance().debug3("After normalizeTheDb, there are "
 			theDbio.deletePlatform(p);
 		theDb.platformList.clear();
 
-		if (!(theDbio instanceof CwmsSqlDatabaseIO))
-		{
-			for(Iterator<Site> sit = theDb.siteList.iterator(); sit.hasNext(); )
-			{
-				Site s = sit.next();
-				theDbio.deleteSite(s);
-			}
-			theDb.siteList.clear();
-		}
-		
+		// NOTE: Never delete SITE records. Too many dependencies outside of DECODES. 
+		// (block of code for removing sites removed 10/26/2022)
 		
 		for(PlatformConfig pc : theDb.platformConfigList.values())
 			theDbio.deleteConfig(pc);
@@ -484,15 +456,18 @@ Logger.instance().debug3("After normalizeTheDb, there are "
 		// Note: EU list and conversions are always assumed to be complete.
 		// Therefore just empty the Java collections. When new ones are read
 		// they will automatically delete all the old ones.
-		theDb.unitConverterSet.clear();
-		theDb.engineeringUnitList.clear();
-
-		// Likewise DataTypes are read/written as a set. Just clear the
-		// Java collection and only the new ones will survive.
-		theDb.dataTypeSet.clear();
-
-		// Likewise again for enumerations.
-		theDb.enumList.clear();
+		if (!platformRelatedOnlyArg.getValue())
+		{
+			theDb.unitConverterSet.clear();
+			theDb.engineeringUnitList.clear();
+	
+			// Likewise DataTypes are read/written as a set. Just clear the
+			// Java collection and only the new ones will survive.
+			theDb.dataTypeSet.clear();
+	
+			// Likewise again for enumerations.
+			theDb.enumList.clear();
+		}
 	}
 
 	/** 
@@ -560,6 +535,25 @@ Logger.instance().debug3("After normalizeTheDb, there are "
 
 			info("Processing '" + s + "'");
 			DatabaseObject ob = null;
+			
+			// If -p argument is used set a filter to skip non-platform-related elements.
+			if (platformRelatedOnlyArg.getValue())
+			{
+				topParser.setElementFilter(
+					new ElementFilter()
+						{
+							@Override
+							public boolean acceptElement(String elementName)
+							{
+								return elementName.equalsIgnoreCase(XmlDbTags.Platform_el)
+									|| elementName.equalsIgnoreCase(XmlDbTags.NetworkList_el)
+									|| elementName.equalsIgnoreCase(XmlDbTags.PlatformConfig_el)
+									|| elementName.equalsIgnoreCase(XmlDbTags.EquipmentModel_el)
+									|| elementName.equalsIgnoreCase(XmlDbTags.Site_el);
+							}
+						});
+			}
+
 			try { ob = topParser.parse(new File(s)); }
 			catch(org.xml.sax.SAXException e)
 			{
@@ -594,26 +588,6 @@ Logger.instance().debug3("After normalizeTheDb, there are "
 				failure("Cannot import PlatformList files! '" + s + "'");
 				throw new DatabaseException("Cannot import PlatformList XML files!");
 			}
-//			else if (ob instanceof EnumList)
-//			{
-//				// XML parser will add any enums it sees into the EnumList
-//				// of the current database (stageDb). So I just need to record
-//				// the fact that I saw an EnumList file.
-//				enumsModified = true;
-//			}
-//			else if (ob instanceof EngineeringUnitList)
-//			{
-//				// XML parser will add any EUs and Converters that it sees to
-//				// the EU List and UC Set of the current database (stageDb). 
-//				// So I just need to record the fact that I saw an EUList File.
-//				euListSeen = true;
-//			}
-			
-//Logger.instance().info("After readXmlFiles...");
-//DataType dt = DataType.getDataType("NRCS", "SRDOX");
-//Logger.instance().info("From DataType.getDatatype: " + dt + ", displayName=" + dt.getDisplayName());
-//dt = stageDb.dataTypeSet.get("NRCS", "SRDOX");
-//Logger.instance().info("From stageDb.dataTypeSet.get: " + dt + ", displayName=" + dt.getDisplayName());
 		}
 
 		/*
@@ -668,10 +642,6 @@ Logger.instance().debug3("After normalizeTheDb, there are "
 	*/
 	private void mergeStageToTheDb()
 	{
-//Logger.instance().debug3("mergeStageToTheDb 0: #EUs=" + theDb.engineeringUnitList.size());
-//Logger.instance().debug3("the db and stage EU lists are "
-//+ (theDb.engineeringUnitList == stageDb.engineeringUnitList ? "" : "NOT") + " the same.");
-
 		Database.setDb(theDb);
 		
 		if (overwriteDb.getValue())
@@ -708,8 +678,8 @@ Logger.instance().debug3("After normalizeTheDb, there are "
 			newDT.setDisplayName(stageDT.getDisplayName());
 		}
 
-Logger.instance().debug3("mergeStageToTheDb 1: #EUs=" + theDb.engineeringUnitList.size());
-Logger.instance().debug3("mergeStageToTheDb 3: #stageEUs=" + stageDb.engineeringUnitList.size());
+		Logger.instance().debug3("mergeStageToTheDb 1: #EUs=" + theDb.engineeringUnitList.size());
+		Logger.instance().debug3("mergeStageToTheDb 3: #stageEUs=" + stageDb.engineeringUnitList.size());
 
 		if (validateOnly)
 		{
@@ -807,7 +777,6 @@ Logger.instance().debug3("mergeStageToTheDb 3: #stageEUs=" + stageDb.engineering
 		//    ==> Import New Platform
 		//    ==> Remove TM from the different existing platform, and if that platform now has
 		//        no remaining TMs, remove it entirely.
-		//TODO Implement above
 		writePlatformList = false;
 		for(Iterator<Platform> it = stageDb.platformList.iterator(); it.hasNext(); )
 		{
